fixed error when geting clean_html
Refactor Markdown handling and editor components

- Removed Tiptap editor component and its dependencies from package.json.
- Introduced MdRenderer component for rendering sanitized Markdown.
- Created utility functions for converting Markdown to HTML and cleaning HTML.
- Updated welcome page to use MdRenderer for displaying Markdown content.
- Replaced Editor component with a simple textarea for Markdown input in the editor page.
- Adjusted layout server files to import routing utilities from the new path.
- Updated Svelte configuration to include utilities path for easier imports.
