+++ package.json
@@ -28,9 +28,6 @@
 		"@tailwindcss/forms": "^0.5.10",
 		"@tailwindcss/typography": "^0.5.16",
 		"@tailwindcss/vite": "^4.1.7",
-		"@tiptap/core": "^2.12.0",
-		"@tiptap/pm": "^2.12.0",
-		"@tiptap/starter-kit": "^2.12.0",
 		"@types/event-calendar__core": "^4.0.0",
 		"@types/jsdom": "^21.1.7",
 		"@types/node": "^22",
