+++ src/app.html
@@ -4,11 +4,10 @@
 		<title>crib.</title>
 		<meta charset="utf-8" />
 		<link rel="icon" href="%sveltekit.assets%/favicon.ico" />
-		<link rel="stylesheet" href="/css/github.css">
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
 		%sveltekit.head%
 	</head>
-	<body data-sveltekit-preload-data="" class="bg-white dark:bg-gray-800 hover display: contents">
+	<body data-sveltekit-preload-data="" class="hover display: contents bg-white dark:bg-gray-800">
 		%sveltekit.body%
 	</body>
 </html>
