+++ src/routes/(app)/[house_id]/+layout.server.ts
@@ -1,5 +1,5 @@
 import type { House, User } from '$schema_types';
-import { getRoute, safeRedirect } from '$server/utils';
+import { getRoute, safeRedirect } from '$utils/routing.utils';
 import { error, redirect } from '@sveltejs/kit';
 import type { LayoutServerLoad } from './$types';
 import { getHouseFromId } from '$server/db/queries/house';
