+++ svelte.config.js
@@ -11,7 +11,8 @@ const config = {
 			$server: 'src/lib/server',
 			$ui: 'src/lib/ui',
 			$schema_types: 'src/lib/schema_types.ts',
-			$env: 'src/lib/server/env.ts'
+			$env: 'src/lib/server/env.ts',
+			$utils: 'src/lib/utils'
 		}
 	},
 	compilerOptions: {
